certificate-authentication-with-msal/requirements.txt
msal==1.4.0
azure.identity==1.3.1
azure.keyvault==4.1.0